+++ cli/cli/kurtosis_cli_version/kurtosis_cli_version.go
@@ -1,6 +1,0 @@
-package kurtosis_cli_version
-
-const (
-	// !!!!!!!! DO NOT UPDATE THIS! It will be updated automatically during release !!!!!!!!!!
-	KurtosisCLIVersion = "0.49.8"
-)