+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/logs_collector_functions/implementations/fluentbit/fluentbit_config_test.go
@@ -14,3 +14,15 @@ func TestGetOutputKurtosisLabelsForLogs(t *testing.T) {
 	fluentbitKurtosisOutputLabels := getOutputKurtosisLabelsForLogs()
 	require.Equal(t, expectedKurtosisFluentbitOutputLabels, fluentbitKurtosisOutputLabels)
 }
+
+func TestGetModifyFilterRulesKurtosisLabels(t *testing.T) {
+
+	expectedKurtosisGUIDDockerLabelRenameFilterRule := "rename com.kurtosistech.guid comKurtosistechGuid"
+	expectedKurtosisContainerTypeDockerLabelRenameFilterRule := "rename com.kurtosistech.guid comKurtosistechGuid"
+	expectedAmountFilterRules := 2
+
+	filterRulesKurtosisLabels := getModifyFilterRulesKurtosisLabels()
+	require.Contains(t, filterRulesKurtosisLabels, expectedKurtosisGUIDDockerLabelRenameFilterRule)
+	require.Contains(t, filterRulesKurtosisLabels, expectedKurtosisContainerTypeDockerLabelRenameFilterRule)
+	require.Equal(t, expectedAmountFilterRules, len(filterRulesKurtosisLabels))
+}