+++ container-engine-lib/lib/backend_impls/docker/object_attributes_provider/object_attributes_provider.go
@@ -18,7 +18,6 @@ const (
 	engineServerNamePrefix   = "kurtosis-engine"
 	logsDatabaseName         = "kurtosis-logs-db"
 	logsCollectorName        = "kurtosis-logs-collector"
-	logsDatabaseMigratorName = logsDatabaseName + "-migrator"
 
 	//We always use the same name because we are going to have only one instance of this volume,
 	//so when the engine is restarted it mounts the same volume with the previous logs