+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/kurtosis_instruction.go
@@ -1,8 +1,18 @@
 package kurtosis_instruction
 
-import "context"
+import (
+	"context"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_validator"
+)
+
+type InstructionPosition struct {
+	line int32
+	col  int32
+}
 
 type KurtosisInstruction interface {
+	GetPositionInOriginalScript() *InstructionPosition
+
 	GetCanonicalInstruction() string
 
 	Execute(ctx context.Context) error
@@ -10,4 +20,15 @@ type KurtosisInstruction interface {
 	// String is only for easy printing in logs and error messages.
 	// Most of the time it will just call GetCanonicalInstruction()
 	String() string
+
+	// ValidateAndUpdateEnvironment validates if the instruction can be applied to an environment, and mutates that
+	// environment to reflect how Kurtosis would look like after this instruction is successfully executed.
+	ValidateAndUpdateEnvironment(environment *startosis_validator.ValidatorEnvironment) error
+}
+
+func NewInstructionPosition(line int32, col int32) *InstructionPosition {
+	return &InstructionPosition{
+		line: line,
+		col:  col,
+	}
 }