+++ core/server/api_container/server/startosis_engine/startosis_modules/git_module_content_provider/git_module_content_provider_test.go
@@ -1,0 +1,43 @@
+package git_module_content_provider
+
+import (
+	"github.com/stretchr/testify/require"
+	"os"
+	"testing"
+)
+
+const (
+	moduleDirRelPath    = "startosis-modules"
+	moduleTmpDirRelPath = "tmp-startosis-modules"
+)
+
+func TestGitModuleProvider_SucceedsForValidModule(t *testing.T) {
+	moduleDir, err := os.MkdirTemp("", moduleDirRelPath)
+	require.Nil(t, err)
+	defer os.RemoveAll(moduleDir)
+	moduleTmpDir, err := os.MkdirTemp("", moduleTmpDirRelPath)
+	require.Nil(t, err)
+	defer os.RemoveAll(moduleTmpDir)
+
+	provider := NewGitModuleContentProvider(moduleDir, moduleTmpDir)
+
+	sampleStartosisModule := "github.com/kurtosis-tech/sample-startosis-load/sample.star"
+	contents, err := provider.GetModuleContents(sampleStartosisModule)
+	require.Nil(t, err)
+	require.Equal(t, "a = \"World!\"\n", contents)
+}
+
+func TestGitModuleProvider_FailsForNonExistentModule(t *testing.T) {
+	moduleDir, err := os.MkdirTemp("", moduleDirRelPath)
+	require.Nil(t, err)
+	defer os.RemoveAll(moduleDir)
+	moduleTmpDir, err := os.MkdirTemp("", moduleTmpDirRelPath)
+	require.Nil(t, err)
+	defer os.RemoveAll(moduleTmpDir)
+
+	provider := NewGitModuleContentProvider(moduleDir, moduleTmpDir)
+	nonExistentModulePath := "github.com/kurtosis-tech/non-existent-startosis-load/sample.star"
+
+	_, err = provider.GetModuleContents(nonExistentModulePath)
+	require.NotNil(t, err)
+}